requirements.txt
Django==3.1
python-dotenv==0.14.0
requests==2.24.0
social-auth-app-django==4.0.0